refractor gameboard.tsx for easyer code reading
